refactoring: add base generator CLI command, move utils to core
