+++ core/src/main/java/org/jboss/sbomer/core/utils/Constants.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.sbomer.utils;
+package org.jboss.sbomer.core.utils;
 
 public class Constants {
